- Вспомогательные функции для тестов перемещены в helpers
